requirements.txt
asgiref==3.2.3
Django==3.0.3
django-crispy-forms==1.8.1
djangorestframework==3.11.0
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.0
